+++ src/app.js
@@ -45,6 +45,7 @@ app.get('/', (req, res) => {
 });
 
 app.use('/', require('./routes/auth'));
+app.use('/diary', require('./routes/diary'));
 
 app.use((req, res) => {
   res.status(404).send('Not Found');